Answer a string option's first byte without building a `&CStr`
The `'langmap'` fix a commit ago was one site of a shape the option
rewrite left all over the tree: 87 calls asked a string option a question
about its **first byte** -- "is it set at all", "is 'selection'
exclusive", "is 'background' dark" -- through the projecting reader,
which has to walk the value to find its NUL before it can hand out a
`&CStr`. `cg_annotate --diff` still put 34 M Ir of `inbench` and 25 M of
`scrbench` in `CStr::from_bytes_until_nul` after the `'langmap'` site was
fixed, and several of these sites are on the draw and charsize paths
(`p_sel` alone is 20 of them, read per screen line in `getvcol` and
`win_line`'s prologue).

All 87 now read `P_XX.first_byte()`, the constant-time accessor, which is
literally upstream's `*p_sel`/`*p_bg`: the C compares the first byte of
the option's own buffer and never measures the string either.

`first_byte` answers `u8` on both halves now -- `StrOpt` for a global and
`LocalOptStr` for a buffer- or window-local field. It was `c_char` on the
local half, which is the type upstream's `char *` dereferences to and the
wrong one here: every caller compares it against a byte literal, so a
`c_char` meant `as u8` or `as c_int != NUL` at the call site. Thirty-odd
of those conversions are gone with it, `buffer::type`'s `ch()` helper
(which existed only to push `b'p'` into a `c_char`) with them, and
`save_file_ff`'s `.cast_unsigned()`.

The sweep was a throwaway script over `scripts/xform.py`; two sites
needed the parentheses back by hand (`dark != P_BG.first_byte() == b'd'`
is not what `dark != p_bg(|v| cstr::first(v) == b'd')` meant), and
`regexp::context`'s two `as ColNr` casts of a comparison became
`ColNr::from`. 38 SAFETY comments over what is now safe code went with
the calls they explained.

No measurement of its own yet -- it lands with the `'foldcolumn'` work
and both are measured together.

+++ crates/nvim/src/buffer/close.rs
@@ -282,7 +282,7 @@ impl Disposition {
         // The caller must take care of NOT deleting/freeing when 'bufhidden'
         // is "hide" (otherwise we could never free or delete a buffer).
         if buffer.terminal.is_null() {
-            match buffer.b_p_bh.first_byte() as u8 {
+            match buffer.b_p_bh.first_byte() {
                 b'd' => (it.del, it.unload) = (true, true),
                 b'w' => (it.del, it.unload, it.wipe) = (true, true, true),
                 b'u' => it.unload = true,
